+++ packages/nitro-protocol/contracts/NitroAdjudicator.sol
@@ -7,23 +7,23 @@ import './Outcome.sol';
 import './AssetHolder.sol';
 
 /**
-  * @dev The NitroAdjudicator contract extends ForceMove and hence inherits all ForceMove methods, and also extends and implements the Adjudicator interface, allowing for a finalized outcome to be pushed to an asset holder.
-*/
+ * @dev The NitroAdjudicator contract extends ForceMove and hence inherits all ForceMove methods, and also extends and implements the Adjudicator interface, allowing for a finalized outcome to be pushed to an asset holder.
+ */
 contract NitroAdjudicator is Adjudicator, ForceMove {
     /**
-    * @notice Allows a finalized channel's outcome to be decoded and one or more AssetOutcomes registered in external Asset Holder contracts.
-    * @dev Allows a finalized channel's outcome to be decoded and one or more AssetOutcomes registered in external Asset Holder contracts.
-    * @param channelId Unique identifier for a state channel
-    * @param turnNumRecord A turnNum that (the adjudicator knows and stores) is supported by a signature from each participant.
-    * @param finalizesAt The unix timestamp when this channel will finalize
-    * @param stateHash The keccak256 of the abi.encode of the State (struct) stored by the adjudicator
-    * @param challengerAddress The address of the participant whom registered the challenge, if any.
-    * @param outcomeBytes The encoded Outcome of this state channel.
-    */
+     * @notice Allows a finalized channel's outcome to be decoded and one or more AssetOutcomes registered in external Asset Holder contracts.
+     * @dev Allows a finalized channel's outcome to be decoded and one or more AssetOutcomes registered in external Asset Holder contracts.
+     * @param channelId Unique identifier for a state channel
+     * @param turnNumRecord A turnNum that (the adjudicator knows and stores) is supported by a signature from each participant.
+     * @param finalizesAt The unix timestamp when this channel will finalize
+     * @param stateHash The keccak256 of the abi.encode of the State (struct) stored by the adjudicator
+     * @param challengerAddress The address of the participant whom registered the challenge, if any.
+     * @param outcomeBytes The encoded Outcome of this state channel.
+     */
     function pushOutcome(
         bytes32 channelId,
-        uint256 turnNumRecord,
-        uint256 finalizesAt,
+        uint48 turnNumRecord,
+        uint48 finalizesAt,
         bytes32 stateHash,
         address challengerAddress,
         bytes memory outcomeBytes
@@ -34,13 +34,7 @@ contract NitroAdjudicator is Adjudicator, ForceMove {
         bytes32 outcomeHash = keccak256(abi.encode(outcomeBytes));
 
         _requireMatchingStorage(
-            ChannelData(
-                turnNumRecord,
-                finalizesAt,
-                stateHash,
-                challengerAddress,
-                outcomeHash
-            ),
+            ChannelData(turnNumRecord, finalizesAt, stateHash, challengerAddress, outcomeHash),
             channelId
         );
 
@@ -54,23 +48,22 @@ contract NitroAdjudicator is Adjudicator, ForceMove {
                 )
             );
         }
-
     }
 
     /**
-    * @notice Allows a finalized channel's outcome to be decoded and transferAll to be triggered in external Asset Holder contracts.
-    * @dev Allows a finalized channel's outcome to be decoded and one or more AssetOutcomes registered in external Asset Holder contracts.
-    * @param channelId Unique identifier for a state channel
-    * @param turnNumRecord A turnNum that (the adjudicator knows and stores) is supported by a signature from each participant.
-    * @param finalizesAt The unix timestamp when this channel will finalize
-    * @param stateHash The keccak256 of the abi.encode of the State (struct) stored by the adjudicator
-    * @param challengerAddress The address of the participant whom registered the challenge, if any.
-    * @param outcomeBytes The encoded Outcome of this state channel.
-    */
+     * @notice Allows a finalized channel's outcome to be decoded and transferAll to be triggered in external Asset Holder contracts.
+     * @dev Allows a finalized channel's outcome to be decoded and one or more AssetOutcomes registered in external Asset Holder contracts.
+     * @param channelId Unique identifier for a state channel
+     * @param turnNumRecord A turnNum that (the adjudicator knows and stores) is supported by a signature from each participant.
+     * @param finalizesAt The unix timestamp when this channel will finalize
+     * @param stateHash The keccak256 of the abi.encode of the State (struct) stored by the adjudicator
+     * @param challengerAddress The address of the participant whom registered the challenge, if any.
+     * @param outcomeBytes The encoded Outcome of this state channel.
+     */
     function pushOutcomeAndTransferAll(
         bytes32 channelId,
-        uint256 turnNumRecord,
-        uint256 finalizesAt,
+        uint48 turnNumRecord,
+        uint48 finalizesAt,
         bytes32 stateHash,
         address challengerAddress,
         bytes memory outcomeBytes
@@ -81,31 +74,24 @@ contract NitroAdjudicator is Adjudicator, ForceMove {
         bytes32 outcomeHash = keccak256(abi.encode(outcomeBytes));
 
         _requireMatchingStorage(
-            ChannelData(
-                turnNumRecord,
-                finalizesAt,
-                stateHash,
-                challengerAddress,
-                outcomeHash
-            ),
+            ChannelData(turnNumRecord, finalizesAt, stateHash, challengerAddress, outcomeHash),
             channelId
         );
 
-        _transferAllFromAllAssetHolders(channelId,outcomeBytes);
-
+        _transferAllFromAllAssetHolders(channelId, outcomeBytes);
     }
 
     /**
-    * @notice Finalizes a channel by providing a finalization proof, allows a finalized channel's outcome to be decoded and transferAll to be triggered in external Asset Holder contracts.
-    * @dev Finalizes a channel by providing a finalization proof, allows a finalized channel's outcome to be decoded and transferAll to be triggered in external Asset Holder contracts.
-    * @param largestTurnNum The largest turn number of the submitted states; will overwrite the stored value of `turnNumRecord`.
-    * @param fixedPart Data describing properties of the state channel that do not change with state updates.
-    * @param appPartHash The keccak256 of the abi.encode of `(challengeDuration, appDefinition, appData)`. Applies to all states in the finalization proof.
-    * @param outcomeBytes abi.encode of an array of Outcome.OutcomeItem structs.
-    * @param numStates The number of states in the finalization proof.
-    * @param whoSignedWhat An array denoting which participant has signed which state: `participant[i]` signed the state with index `whoSignedWhat[i]`.
-    * @param sigs An array of signatures that support the state with the `largestTurnNum`.
-    */
+     * @notice Finalizes a channel by providing a finalization proof, allows a finalized channel's outcome to be decoded and transferAll to be triggered in external Asset Holder contracts.
+     * @dev Finalizes a channel by providing a finalization proof, allows a finalized channel's outcome to be decoded and transferAll to be triggered in external Asset Holder contracts.
+     * @param largestTurnNum The largest turn number of the submitted states; will overwrite the stored value of `turnNumRecord`.
+     * @param fixedPart Data describing properties of the state channel that do not change with state updates.
+     * @param appPartHash The keccak256 of the abi.encode of `(challengeDuration, appDefinition, appData)`. Applies to all states in the finalization proof.
+     * @param outcomeBytes abi.encode of an array of Outcome.OutcomeItem structs.
+     * @param numStates The number of states in the finalization proof.
+     * @param whoSignedWhat An array denoting which participant has signed which state: `participant[i]` signed the state with index `whoSignedWhat[i]`.
+     * @param sigs An array of signatures that support the state with the `largestTurnNum`.
+     */
     function concludePushOutcomeAndTransferAll(
         uint48 largestTurnNum,
         FixedPart memory fixedPart,
@@ -124,7 +110,7 @@ contract NitroAdjudicator is Adjudicator, ForceMove {
         bytes32[] memory stateHashes = new bytes32[](numStates);
 
         bytes32 outcomeHash = keccak256(abi.encode(outcomeBytes));
-        for (uint256 i = 0; i < numStates; i++) {
+        for (uint48 i = 0; i < numStates; i++) {
             stateHashes[i] = keccak256(
                 abi.encode(
                     State(
@@ -150,38 +136,39 @@ contract NitroAdjudicator is Adjudicator, ForceMove {
         );
 
         // effects
-        
 
         channelStorageHashes[channelId] = _hashChannelData(
-            ChannelData(0, now, bytes32(0), address(0), outcomeHash)
+            ChannelData(0, uint48(now), bytes32(0), address(0), outcomeHash)
         );
         emit Concluded(channelId);
 
         _transferAllFromAllAssetHolders(channelId, outcomeBytes);
     }
 
-
     /**
-    * @notice Triggers transferAll in all external Asset Holder contracts specified in a given outcome for a given channelId.
-    * @dev Triggers transferAll in  all external Asset Holder contracts specified in a given outcome for a given channelId.
-    * @param channelId Unique identifier for a state channel
-    * @param outcomeBytes abi.encode of an array of Outcome.OutcomeItem structs.
-    */
-    function _transferAllFromAllAssetHolders(
-        bytes32 channelId,
-        bytes memory outcomeBytes
-    ) internal {
-
+     * @notice Triggers transferAll in all external Asset Holder contracts specified in a given outcome for a given channelId.
+     * @dev Triggers transferAll in  all external Asset Holder contracts specified in a given outcome for a given channelId.
+     * @param channelId Unique identifier for a state channel
+     * @param outcomeBytes abi.encode of an array of Outcome.OutcomeItem structs.
+     */
+    function _transferAllFromAllAssetHolders(bytes32 channelId, bytes memory outcomeBytes)
+        internal
+    {
         Outcome.OutcomeItem[] memory outcome = abi.decode(outcomeBytes, (Outcome.OutcomeItem[]));
 
         for (uint256 i = 0; i < outcome.length; i++) {
-            Outcome.AssetOutcome memory assetOutcome = abi.decode(outcome[i].assetOutcomeBytes,(Outcome.AssetOutcome));
+            Outcome.AssetOutcome memory assetOutcome = abi.decode(
+                outcome[i].assetOutcomeBytes,
+                (Outcome.AssetOutcome)
+            );
             if (assetOutcome.assetOutcomeType == uint8(Outcome.AssetOutcomeType.Allocation)) {
-                AssetHolder(outcome[i].assetHolderAddress).transferAllAdjudicatorOnly(channelId,assetOutcome.allocationOrGuaranteeBytes);
+                AssetHolder(outcome[i].assetHolderAddress).transferAllAdjudicatorOnly(
+                    channelId,
+                    assetOutcome.allocationOrGuaranteeBytes
+                );
             } else {
                 revert();
             }
-
         }
     }
 
@@ -200,15 +187,9 @@ contract NitroAdjudicator is Adjudicator, ForceMove {
         uint256 nParticipants,
         bool[2] memory isFinalAB, // [a.isFinal, b.isFinal]
         ForceMoveApp.VariablePart[2] memory ab, // [a,b]
-        uint256 turnNumB,
+        uint48 turnNumB,
         address appDefinition
-        ) public pure returns (bool) {
-
+    ) public pure returns (bool) {
         return _requireValidTransition(nParticipants, isFinalAB, ab, turnNumB, appDefinition);
-
     }
 }
-
-
-
-
